% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spag_Lrev.R
\name{.mixed_formula}
\alias{.mixed_formula}
\title{Build the linear mixed-effects model formula}
\usage{
.mixed_formula(df, x, time, ID)
}
\arguments{
\item{df}{A data frame containing the longitudinal data.}

\item{x}{The name of the response variable, provided as a character string.}

\item{time}{The name of the time variable, provided as a character string.}

\item{ID}{The name of the subject identifier variable,
provided as a character string.}
}
\value{
A character string containing the model formula.
}
\description{
Internal function used by \code{\link[=spag_Lrev]{spag_Lrev()}} to construct the formula
used to calculate longitudinal residuals.
}
\keyword{internal}
