% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{longi_demo}
\alias{longi_demo}
\title{Example longitudinal data for Lrev}
\format{
A data frame with observations collected repeatedly for
multiple individuals:
\describe{
\item{ID}{Individual identifier.}
\item{Gender}{Sex of the individual.}
\item{Center}{Center of treatment}
\item{Treatment}{Treatment group.}
\item{Age}{Age in years.}
\item{Time}{Time of measurement in days}
\item{Score}{Clinical score with multiple measures per individual}
\item{Visit}{Discrete time variable}
}
}
\source{
Synthetic data generated for demonstration purposes.
}
\usage{
longi_demo
}
\description{
A synthetic longitudinal dataset used to demonstrate the longitudinal
data review application \code{\link[=shiny_Lrev]{shiny_Lrev()}} and the plotting function
\code{\link[=spag_Lrev]{spag_Lrev()}}.
}
\keyword{datasets}
