% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spag_Lrev.R
\name{spag_Lrev}
\alias{spag_Lrev}
\title{Spaghetti plot for longitudinal data review}
\usage{
spag_Lrev(
  df,
  x,
  time,
  ID,
  group = "Aucune",
  incl = NULL,
  isol = "Aucune",
  y_type = "Valeurs brutes",
  ylim = NULL,
  xlim = NULL,
  alpha = 0.7,
  base_color = ggplot2::scale_color_grey(start = 0.2, end = 0.8),
  group_color = ggplot2::scale_color_viridis_d(),
  text_size = 11
)
}
\arguments{
\item{df}{A data frame containing the longitudinal data, meaning 1 line = 1 observation.}

\item{x}{The name of the variable to be displayed on the y-axis,
provided as a character string.}

\item{time}{The name of the time variable, provided as a character string. The time variable should be either a numeric variable or an ordered factor.}

\item{ID}{The name of the subject identifier variable,
provided as a character string.}

\item{group}{An optional qualitative grouping variable, provided as a
character string. Defaults to \code{"Aucune"} (none).}

\item{incl}{An optional character vector specifying the levels of \code{group}
to include in the plot. Defaults to \code{NULL}.}

\item{isol}{An optional character vector specifying the IDs of subjects
to isolate. Defaults to \code{"Aucune"} (none).}

\item{y_type}{Specifies whether to display \code{"Valeurs brutes"} (raw values)
or \code{"Résidus"} (residuals from a linear mixed-effects model). The mixed model is specified after the filtering by \code{incl} and before the filtering by \code{isol} but is not stratified by the \code{group}.
If \code{time} is a numeric variable, random effects are applied to the intercept and to \code{time}. If time is an ordered factor only a random intercept is applied.
Defaults to \code{"Valeurs brutes"}.}

\item{ylim}{An optional numeric vector of length two specifying the limits
of the y-axis. Only applied when \code{y_type = "Valeurs brutes"}.}

\item{xlim}{An optional numeric vector of length two specifying the limits
of the x-axis. Only applied when \code{time} is numeric.}

\item{alpha}{The transparency of the points and trajectories.
Defaults to \code{0.7}.}

\item{base_color}{A ggplot2 color scale used to customize the color palette
when no grouping variable is specified. Defaults to
\code{\link[ggplot2:scale_grey]{ggplot2::scale_color_grey()}} with start = \code{0.2}, end = \code{0.8}.}

\item{group_color}{A ggplot2 color scale used to customize the color palette
when a grouping variable is specified. Defaults to
\code{\link[ggplot2:scale_viridis]{ggplot2::scale_color_viridis_d()}}.}

\item{text_size}{The base text size of the ggplot2 theme.
Defaults to \code{11}.}
}
\value{
A ggplot2 object.
}
\description{
Creates a longitudinal spaghetti plot showing individual trajectories
over time. Observations can be displayed as raw values or as residuals
from a linear mixed-effects model. Individuals can optionally be
selected or filtered according to a grouping variable.
}
\examples{
# Visualise individual trajectories over time
spag_Lrev(
  df = longi_demo,
  x = "Score",
  time = "Time",
  ID = "ID"
)

# Also works with discrete time variables
spag_Lrev(
  df = longi_demo,
  x = "Score",
  time = "Visit",
  ID = "ID"
)

# Visualise trajectories by center and focus on a plot area
spag_Lrev(
  df = longi_demo,
  x = "Score",
  time = "Time",
  ID = "ID",
  group = "Center",
  xlim=c(90,180),
  ylim=c(10,75)
)

# Display residuals from a linear mixed-effects model
spag_Lrev(
  df = longi_demo,
  x = "Score",
  time = "Time",
  ID = "ID",
  y_type = "R\u00e9sidus"
)
}
